feat: add version command and allow config path
This commit add a new command to the app in the form of the `version`
command that allows to easily see what version of the package we're
using. Previously this was only visible when the server was starting.

It also add a new parameter that would allow the user to specify the
configuration folder to use. The default value is still `conf/`, but now
we can also specify the path we want.